Rename components to match new import system
