+++ galaxyui/src/app/react/containers/add-content-container.tsx
